+++ src/components/CalculateNodeCount.tsx
@@ -1,6 +1,6 @@
 import { useContext, useEffect } from 'react';
 import { Skill } from 'models';
-import SkillTreeContext from '../context/SkillTreeContext';
+import SkillContext from '../context/SkillContext';
 
 interface Props {
   data: Skill[];
@@ -17,7 +17,7 @@ function calculateNodeCount(data: Skill[]): number {
 }
 
 function CalculateNodeCount({ data }: Props) {
-  const { addToSkillCount } = useContext(SkillTreeContext);
+  const { addToSkillCount } = useContext(SkillContext);
 
   useEffect(() => {
     const count = calculateNodeCount(data);
