+++ src/components/__tests__/SkillTree.test.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { render, fireEvent, act } from '@testing-library/react';
 import SkillTree from '../SkillTree';
 import MockLocalStorage from '../__mocks__/mockLocalStorage';
-import { SkillProvider } from '../../context/SkillAppContext';
+import { SkillProvider } from '../../context/AppContext';
 import SkillTreeGroup from '../../components/SkillTreeGroup';
-import { SkillTreeProvider } from '../../context/SkillTreeContext';
+import { SkillTreeProvider } from '../../context/SkillContext';
 
 const mockSkillTreeData = [
   {
