+++ src/index.tsx
@@ -1,4 +1,4 @@
 export { default as SkillTreeGroup } from './components/SkillTreeGroup';
 export { default as SkillTree } from './components/SkillTree';
-export { SkillProvider } from './context/SkillAppContext';
+export { SkillProvider } from './context/AppContext';
 export { Skill as SkillType } from './models/index';
